requirements.txt
discord.py==1.3.3
requests==2.23.0